% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.array-pkg.R
\docType{package}
\name{list.array}
\alias{list.array}
\alias{list.array-package}
\title{\pkg{list.array} - Array of list objects}
\description{
A 2d array with lists as elements.
}
\author{
Mark Heckmann
}
